Changed "define" to "Notice"
The existing text instructed users to define the header. However, when you import a postman collection the header would already have been defined. It doesn't make sense to tell users to define something that is already defined. So, I changed the sentence to "Notice that the headers.....are already defined.